// projekat/frontend/src/app/model/privatnaBasta.ts
export class PrivatnaBasta{

    datum_vreme: Date;
    kvadratura_baste: number;
    kvadratura_bazen: number;
    kvadratura_zelenilo: number;
    kvadratura_inventar: number;
    datumZakazivanja: Date;
    imeFirme: string;
    brBazena: string;
    renoviranje: boolean;
    obradjen: boolean;
    procenaZavrsetka: Date;
    potvrdaOdrzavanja: boolean;

    
}